api/using-gpu/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.2.0
fastapi==0.109.0
gTTS==2.5.0
h11==0.14.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
joblib==1.3.2
moviepy==1.0.3
numpy==1.26.3
pillow==10.2.0
proglog==0.1.10
pydantic==2.5.3
pydantic-core==2.14.6
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
threadpoolctl==3.2.0
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
transformers
tensorflow